refactor(meta): lift .cargo_vcs_info.json parser into a testable src helper
The `.cargo_vcs_info.json` hand-parse was the fiddliest new code and lived
in build.rs, where it can't be reached by `cargo test` (build scripts are
not in the test harness). Move the pure parsing into
`src/meta/vcs_info.rs` and share it with build.rs via `include!`, so the
logic has one home and gains unit tests (real vcs-info, pretty-printed
whitespace, and malformed inputs that must degrade to `unknown`).

Plain `//` comments (not `//!`) in that file on purpose: build.rs
`include!`s it mid-module, where an inner doc comment is a compile error.
Also document that the 7-char pin is git's default `--short` minimum and
may be one char shorter than an auto-lengthened git-checkout build's hash
(purely cosmetic; `-V` is not byte-gated).

Addresses review items 3 & 4 on PR #1073.
